Use unsafe mul algo for BcMath and remove UnsafePoint
